requirements.txt
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
msal>=1.27.0
PyYAML>=6.0
flask>=3.0.0
tqdm>=4.0.0
